+++ src/main/scala/io/github/facaiy/dag/parallel/Implicits.scala
@@ -1,15 +1,19 @@
-package io.github.facaiy.DAG.parallel
+package io.github.facaiy.dag.parallel
 
 import scala.concurrent.{Await, ExecutionContext, Future}
 
-import io.github.facaiy.DAG.{DAG => ParentDAG}
-import io.github.facaiy.DAG.core.{DAGNode, InputNode, InternalNode, LazyCell}
-import io.github.facaiy.DAG.serial.DAG.Nodes
+import io.github.facaiy.dag.core.{DAGNode, InputNode, InternalNode, LazyCell}
 
 /**
  * Created by facai on 6/2/17.
  */
-object DAG extends ParentDAG {
+object Implicits {
+  import scala.language.implicitConversions
+
+  implicit def asFutureCell[K, V](nodes: Seq[DAGNode[K, V]]): FutureCell[K, V] = FutureCell(nodes)
+
+  implicit def asLazyFuture[A](lc: LazyCell[Future[A]]): LazyFuture[A] = LazyFuture(lc)
+
   case class FutureCell[K, V](nodes: Seq[DAGNode[K, V]]) {
     def toParallel(implicit executor: ExecutionContext): Seq[DAGNode[K, Future[V]]] = {
 
@@ -36,17 +40,4 @@ object DAG extends ParentDAG {
     def getValue(timeOut: Int): A =
       Await.result(lc.get(), timeOut.seconds)
   }
-}
-
-object Implicits {
-  import DAG._
-
-  import scala.language.implicitConversions
-
-  // TODO(facai), duplication of serial.DAG.asNodesOps
-  implicit def asNodes[K, V](nodes: Seq[DAGNode[K, V]]): Nodes[K, V] = Nodes(nodes)
-
-  implicit def asFutureCell[K, V](nodes: Seq[DAGNode[K, V]]): FutureCell[K, V] = FutureCell(nodes)
-
-  implicit def asLazyFuture[A](lc: LazyCell[Future[A]]): LazyFuture[A] = LazyFuture(lc)
 }
